pyproject.toml
[tool.poetry]
name = "assume_framework"
version = "0.0.1"
description = "ASSUME - Agent based energy market modeling using mango agents"
authors = ["ASSUME developers <nick.harder@inatech.uni-freiburg.de>"]
license = "LICENSE"
readme = "README.md"

homepage = "https://assume.readthedocs.io"
repository = "https://github.com/assume-framework/assume"

keywords = ["agent based simulation", "energy market", "reinforcement learning", "market simulation", "simulation"]

classifiers=[
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 (AGPLv3)",
    "Operating System :: OS Independent",
    "Framework :: asyncio",
    "Framework :: mango-agents",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [
    { include="assume", from="." },
]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.5.1"
mango-agents = {git = "https://gitlab.com/maurerle/mango"}
tqdm = "^4.64.1"
python-dateutil = "^2.8.2"
sqlalchemy = "^2.0.9"
pandas = {version = "^2.0.0"}
psycopg2-binary = "^2.9.5"
pyyaml = "^6.0"
nest-asyncio = "^1.5.6"
black = {version = "^23.3.0", optional = true}
isort = {version = "^5.12.0", optional = true}
mypy = {version = "^1.1.1", optional = true}
matplotlib = {version = "^3.7.2", optional = true}
pytest = {version = "^7.2.2", optional = true}
pytest-cov = {version = "^4.1.0", optional = true}
pytest-asyncio = {version = "^0.21.1", optional = true}
torch = {version = "^2.0.1", optional = true}
pyomo = {version = "^6.6.1", optional = true}
glpk = {version = "^0.4.7", optional = true}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.1.1"
pytest = "^7.2.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"


[tool.poetry.extras]
learning = ["torch"]
optimize = ["pyomo", "glpk"]
test = ["black", "isort", "matplotlib", "pytest", "pytest-cov", "pytest-asyncio", "pyomo", "glpk"]


[tool.poetry.scripts]
assume = "assume.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest]
testpaths = "tests"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore:Call to deprecated create function FieldDescriptor",
    "ignore:Call to deprecated create function Descriptor",
    "ignore:Call to deprecated create function FileDescriptor",
    "ignore:Call to deprecated create function EnumValueDescriptor",
    "ignore:Call to deprecated create function EnumDescriptor",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "require_learning",
]
addopts = "--ignore=assume-db"

[tool.coverage.run]
omit = ["tests/*"]
source = ["assume"]

[tool.coverage.report]
# https://stackoverflow.com/a/5850364
exclude_lines = ["if __name__ == .__main__.:"]

[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
# output
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true
ignore_missing_imports = true
